updating db initialiser

+++ CameraShop/Models/PaymentCard.cs
@@ -5,7 +5,7 @@ namespace CameraShop.Models
     public class PaymentCard
     {
 
-        public int CardID { get; set; }
+        public long CardID { get; set; }
         
         [Required, StringLength(20)]
         public string CardName { get; set; }
